InzOprProject/requirements.txt
adjustText==0.7.3
asgiref==3.4.1
blank==0.1.0
cycler==0.11.0
Django==4.0.1
djangorestframework==3.13.1
Employee==0.0.4
fonttools==4.29.0
future==0.18.2
iso8601==1.0.2
joblib==1.1.0
kiwisolver==1.3.2
lark-parser==0.12.0
many==0.6.9
matplotlib==3.5.1
numpy==1.22.1
packaging==21.3
pandas==1.4.0
patsy==0.5.2
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
required==0.4.0
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
serial==0.0.97
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
statsmodels==0.13.1
threadpoolctl==3.0.0
tqdm==4.62.3
